Move `Address`, `ButtonLink` and `CustomerInfo` to `/react`
