+++ react/CustomerInfo.tsx
